Display factR2 samples and expression data

If expression data from multiple samples is provided, factRObject-class will store the expression values at the gene, transcript and alternative exon levels as well as samples metadata. The functions below display contents pertaining to expression and samples data:

# S4 method for factR
rowNames(object)

# S4 method for factR
samples(object)

# S4 method for factR
colNames(object)

# S4 method for factR
colNames(object) <- value

# S4 method for factR
ident(object)

# S4 method for factR
counts(object, ..., set = NULL, slot = "data")

Arguments

object

factRObject

value

New sample names. Length should match number of columns

...

One or more features to display. Can be the following:

gene_id: ID of gene to plot

gene_name: Name of gene to plot

transcript_id: ID of transcript to plot

set

Set metadata to display. Can be "gene", "transcript" or "AS".

slot

Data slot to display. Can be "counts" or "data" (default)

Value

`counts`: Count matrix

All other functions: Character vector of values

Details

See factR-exp for example codes

See also