Expression/samples-related functions

A set of functions to incorporate expression data into factR object and tests regulatory potential of alternative splicing events.

Correlates exon inclusion levels with gene expression levels

# S4 method for factR
addTxCounts(object, countData, sampleData = NULL, psi = NULL, verbose = TRUE)

# S4 method for factR
testGeneCorr(object, vst = TRUE, min_n = 3, ...)

Arguments

object

factR object

countData

Can be one of the following:

Path to local count matrix file in .tsv or .csv format

Matrix object containing transcript-level expression counts data.

sampleData

(Optional) Dataframe containing samples information. Dataframe rows can be named to match the sample names in `countData`. If `sampleData` has no row names, function will attempt to pick the column containing sample names and assign it as rownames.

psi

(Optional) Exon-level splicing inclusion data. Can be one of the following:

Path to local file in .tsv or .csv format

Matrix object

Rownames of matrix should be in chr:start-end format

verbose

Boolean value as to whether messages should be printed (Default: TRUE)

vst

whether to apply variance stabilization on splicing and expression levels

min_n

minimum number of non-NA samples required for correlation testing

...

additional arguments parsed to cor.test function

Value

factRObject with updated counts data and samples metadata.

factRObject with updated ASE metadata

Details

`addTxCounts` converts transcript-level expression counts into gene-level expression counts. If "psiData" is not provided, this function will determine the inclusion level of alternative splicing events (psi) for each sample by calculating the proportion of transcripts containing the splicing event over all transcripts that spans that splicing event. If "psiData" is provided, will check for common exon coordinates and non-overlapping ones detected by factR2 will be set to NA. Row names of "sampleData" dataframe can be unnamed, and this function will search for column variable corresponding to sample names. If unsuccessful or if row names of "sampleData" does not match to column names of "countData", an error will be returned.

See also

Examples

## get path to sample GTF and expression data
gtf <- system.file("extdata", "pb_custom.gtf.gz", package = "factR2")
counts <- system.file("extdata", "pb_expression.tsv.gz", package = "factR2")

## create factRObject with expression counts
factRObject <- createfactRObject(gtf, "vM33", countData = counts)
#> 🡆  Checking inputs
#> 🡆  Checking factRObject
#> 🡆  Adding custom transcriptome
#>     ℹ Importing from local directory
#> 🡆  Adding annotation
#>     ℹ Importing from URL
#> 🡆  Adding genome sequence
#>     ℹ Using BSgenome object
#> Warning: 4 seqlevel(s) in ``your custom GTF`` are not found in ``reference annotation``
#> 🡆  Matching chromosome names
#> 🡆  Matching gene information
#>     Number of mismatched gene_ids found: 83520
#>     --> Attempting to match ensembl gene_ids...
#>     --> No ensembl gene ids found in query
#>     ---> Attempting to match gene_ids by finding overlapping coordinates...
#>     ---> 65203 gene_id matched
#>     Total gene_ids corrected: 65203
#>     Remaining number of mismatched gene_ids: 18317
#> 🡆  Creating factRset objects
#>     ℹ Adding gene information
#>     ℹ Adding transcript information
#>     ℹ Adding alternative splicing information
#> 🡆  Annotating novel transcripts
#> 🡆  Annotating novel AS events
#> 🡆  Adding expression data
#>     ℹ Importing from local file
#> 🡆  Creating samples metadata
#> 🡆  Processing expression data
#>     ℹ Adding gene counts
#>     ℹ Adding spliced-event counts
#>     ℹ Normalizing counts
#> ℹ factRobject created!

### This can also be done post-creation of a factR object
if (FALSE) {
factRObject <- createfactRObject(gtf, "vM33")
factRObject <- addTxCounts(factRObject, counts)}

## access counts data
counts(factRObject)  # returns normalised expression data from current Set
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("counts"), generic = structure("counts", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "counts"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function ‘counts’ for signature ‘"factR"’
counts(factRObject, "Ptbp1")  # same as above, but only for specific genes
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("counts"), generic = structure("counts", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "counts"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function ‘counts’ for signature ‘"factR"’
counts(factRObject, "Ptbp1", set = "transcript")  # get transcript-level expression
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("counts"), generic = structure("counts", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "counts"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function ‘counts’ for signature ‘"factR"’
counts(factRObject, "Ptbp1", set = "gene") # get gene-level expression
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("counts"), generic = structure("counts", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "counts"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function ‘counts’ for signature ‘"factR"’
counts(factRObject, "Ptbp1", set = "gene", slot = "counts") # get gene-level count
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (object, ...)     standardGeneric("counts"), generic = structure("counts", package = "BiocGenerics"),     package = "BiocGenerics", group = list(), valueClass = character(0),     signature = "object", default = NULL, skeleton = (function (object,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "counts"), domain = NA))(object, ...)), <environment>): unable to find an inherited method for function ‘counts’ for signature ‘"factR"’

## access samples metadata
samples(factRObject)
#>            proj.ident
#> P14_M5_1 factRProject
#> P14_M5_2 factRProject
#> P14_M5_3 factRProject
#> P14_M6_1 factRProject
#> P14_M6_2 factRProject
#> P14_M6_3 factRProject
#> P28_M1_1 factRProject
#> P28_M1_2 factRProject
#> P28_M1_3 factRProject
#> P28_M2_1 factRProject
#> P28_M2_2 factRProject
#> P28_M2_3 factRProject
ident(factRObject)   # prints out the current identity
#> [1] "proj.ident"

## run correlation between gene expression and exon inclusion
factRObject <- testGeneCorr(factRObject)
#> Error in dplyr::select(., AS_id, gene_id, ASNMDtype): Can't subset columns that don't exist.
#> ✖ Column `ASNMDtype` doesn't exist.
ase(factRObject)
#> ℹ Set `show_more to TRUE to show more info`
#> # A tibble: 48,197 × 8
#>    AS_id   gene_id               gene_name coord       AStype strand width novel
#>    <chr>   <chr>                 <chr>     <chr>       <fct>  <fct>  <int> <chr>
#>  1 AS01129 ENSMUSG00000033813.16 Tcea1     chr1:49570… AD     +        231 yes  
#>  2 AS01642 ENSMUSG00000025907.15 Rb1cc1    chr1:62851… AD     +       1104 yes  
#>  3 AS01647 ENSMUSG00000025907.15 Rb1cc1    chr1:62929… AF     +        499 yes  
#>  4 AS01649 ENSMUSG00000025907.15 Rb1cc1    chr1:63043… RI     +        141 yes  
#>  5 AS01650 ENSMUSG00000025907.15 Rb1cc1    chr1:63152… RI     +        178 yes  
#>  6 AS01651 ENSMUSG00000025907.15 Rb1cc1    chr1:63155… RI     +         80 yes  
#>  7 AS01652 ENSMUSG00000025907.15 Rb1cc1    chr1:63198… RI     +         87 yes  
#>  8 AS01657 ENSMUSG00000025907.15 Rb1cc1    chr1:63313… AD     +        749 yes  
#>  9 AS01658 ENSMUSG00000025907.15 Rb1cc1    chr1:63331… RI     +         81 yes  
#> 10 AS01659 ENSMUSG00000025907.15 Rb1cc1    chr1:63334… AD     +        173 yes  
#> # ℹ 48,187 more rows