Display factR2 object data

A factRObject-class contains different types of data at the gene, transcript, alternative splicing (AS) and protein domain levels. The functions below are designed to display specific contents of a factRObject.

# S4 method for factR
granges(object, ..., set = NULL)

# S4 method for factR
activeSet(object)

# S4 method for factR
activeSet(object) <- value

# S4 method for factR
listSets(object)

# S4 method for factR
features(object, ..., set = NULL, show_more = FALSE)

# S4 method for factR
genes(object, ..., show_more = FALSE)

# S4 method for factR
gns(object, ..., show_more = FALSE)

# S4 method for factR
transcripts(object, ..., show_more = FALSE)

# S4 method for factR
txs(object, ..., show_more = FALSE)

# S4 method for factR
ase(object, ..., show_more = FALSE)

# S4 method for factR
domains(object, ...)

Arguments

object

factRObject

...

One or more features to display. Can be the following:

gene_id: ID of gene to plot

gene_name: Name of gene to plot

transcript_id: ID of transcript to plot

set

Set metadata to display. Can be "gene", "transcript" or "AS".

value

Character value of one of the following: "gene", "transcript" or "AS"

Value

`granges`: GenomicRanges object of selected features

`activeSet` and `listSets`: Character value/vector

All other functions: Tibble dataframe containing metadata of selected features

Examples

### Load sample factRObject
data("factRsample")

## Prints out activeSet
activeSet(factRsample)
#> [1] "AS"

## Change activeSet
activeSet(factRsample) <- "transcript"

## Returns coordinates and metadata of features as a GenomicRanges object
granges(factRsample)   # from activeSet
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (x, use.names = TRUE, use.mcols = FALSE,         ...)     standardGeneric("granges"), generic = structure("granges", package = "GenomicRanges"),     package = "GenomicRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         use.names = TRUE, use.mcols = FALSE, ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "granges"), domain = NA))(x, use.names, use.mcols, ...)),     <environment>): unable to find an inherited method for function ‘granges’ for signature ‘"factR"’
granges(factRsample, "Dab2")   # specific features
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (x, use.names = TRUE, use.mcols = FALSE,         ...)     standardGeneric("granges"), generic = structure("granges", package = "GenomicRanges"),     package = "GenomicRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         use.names = TRUE, use.mcols = FALSE, ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "granges"), domain = NA))(x, use.names, use.mcols, ...)),     <environment>): unable to find an inherited method for function ‘granges’ for signature ‘"factR"’
granges(factRsample, "Dab2", set = "gene")   # specific features from different Set
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list(structure("factR", package = "factR2")), new("standardGeneric",     .Data = function (x, use.names = TRUE, use.mcols = FALSE,         ...)     standardGeneric("granges"), generic = structure("granges", package = "GenomicRanges"),     package = "GenomicRanges", group = list(), valueClass = character(0),     signature = "x", default = NULL, skeleton = (function (x,         use.names = TRUE, use.mcols = FALSE, ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "granges"), domain = NA))(x, use.names, use.mcols, ...)),     <environment>): unable to find an inherited method for function ‘granges’ for signature ‘"factR"’

## Returns metadata of features
features(factRsample)   # from activeSet
#> ℹ Set `show_more to TRUE to show more info`
#> # A tibble: 154 × 8
#>    transcript_id         gene_id        gene_name strand width novel cds   nmd  
#>    <chr>                 <chr>          <chr>     <fct>  <int> <chr> <chr> <chr>
#>  1 ENSMUST00000022746.12 ENSMUSG000000… Osmr      -       5491 no    no    no   
#>  2 ENSMUST00000175862.7  ENSMUSG000000… Osmr      -       2837 no    no    no   
#>  3 ENSMUST00000176554.1  ENSMUSG000000… Osmr      -        908 no    no    no   
#>  4 ENSMUST00000176826.1  ENSMUSG000000… Osmr      -       4028 no    no    no   
#>  5 ENSMUST00000177263.1  ENSMUSG000000… Osmr      -       6160 no    no    no   
#>  6 ENSMUST00000177478.1  ENSMUSG000000… Osmr      -       1727 no    no    no   
#>  7 MSTRG.14573.1         ENSMUSG000000… Osmr      *        877 yes   no    no   
#>  8 ENSMUST00000090461.11 ENSMUSG000000… Fyb       +       2678 no    no    no   
#>  9 ENSMUST00000159604.1  ENSMUSG000000… Fyb       +        771 no    no    no   
#> 10 ENSMUST00000159714.1  ENSMUSG000000… Fyb       +       2170 no    no    no   
#> # ℹ 144 more rows
features(factRsample, "Dab2")   # specific features
#> ℹ Set `show_more to TRUE to show more info`
#> # A tibble: 16 × 8
#>    transcript_id         gene_id        gene_name strand width novel cds   nmd  
#>    <chr>                 <chr>          <chr>     <fct>  <int> <chr> <chr> <chr>
#>  1 ENSMUST00000078019.12 ENSMUSG000000… Dab2      +       3793 no    no    no   
#>  2 ENSMUST00000080880.11 ENSMUSG000000… Dab2      +       4670 no    no    no   
#>  3 ENSMUST00000110663.8  ENSMUSG000000… Dab2      +       3838 no    no    no   
#>  4 ENSMUST00000110664.8  ENSMUSG000000… Dab2      +       4457 no    no    no   
#>  5 ENSMUST00000159046.1  ENSMUSG000000… Dab2      +        438 no    no    no   
#>  6 ENSMUST00000159490.7  ENSMUSG000000… Dab2      +        444 no    no    no   
#>  7 ENSMUST00000159552.1  ENSMUSG000000… Dab2      +        922 no    no    no   
#>  8 ENSMUST00000160134.7  ENSMUSG000000… Dab2      +       1460 no    no    no   
#>  9 ENSMUST00000161040.7  ENSMUSG000000… Dab2      +       1184 no    no    no   
#> 10 ENSMUST00000161558.1  ENSMUSG000000… Dab2      +       2119 no    no    no   
#> 11 ENSMUST00000161812.7  ENSMUSG000000… Dab2      +       3774 no    no    no   
#> 12 ENSMUST00000162094.2  ENSMUSG000000… Dab2      +        737 no    no    no   
#> 13 ENSMUST00000162140.1  ENSMUSG000000… Dab2      +       1406 no    no    no   
#> 14 ENSMUST00000163082.1  ENSMUSG000000… Dab2      +       3839 no    no    no   
#> 15 MSTRG.14564.1         ENSMUSG000000… Dab2      *        213 yes   no    no   
#> 16 MSTRG.14566.1         ENSMUSG000000… Dab2      *        202 yes   no    no   
features(factRsample, "Dab2", set = "gene")   # specific features from different Set
#> ℹ Set `show_more to TRUE to show more info`
#> # A tibble: 1 × 5
#>   gene_id               gene_name strand  width match_level
#>   <chr>                 <chr>     <fct>   <int>       <dbl>
#> 1 ENSMUSG00000022150.16 Dab2      +      140925           4

### This is the same as:
genes(factRsample, "Dab2")
#> ℹ Set `show_more to TRUE to show more info`
#> # A tibble: 1 × 5
#>   gene_id               gene_name strand  width match_level
#>   <chr>                 <chr>     <fct>   <int>       <dbl>
#> 1 ENSMUSG00000022150.16 Dab2      +      140925           4

## To return protein-coding domains, the protein-coding domains need to be predicted first:
factRsample <- buildCDS(factRsample)
factRsample <- getAAsequence(factRsample)
factRsample <- predictDomains(factRsample, "Dab2")
#> ℹ Set `show_more to TRUE to show more info`
#> Warning: Skipped 6 non-coding transcripts

## Then, the domains of the selected gene can be printed as such:
domains(factRsample, "Dab2")
#> # A tibble: 8 × 6
#>   type   description    eval    begin   end transcript_id        
#>   <chr>  <chr>          <chr>   <dbl> <dbl> <chr>                
#> 1 DOMAIN PH domain-like 2.7e-42    33   180 ENSMUST00000078019.12
#> 2 DOMAIN PH domain-like 4.9e-42    33   180 ENSMUST00000080880.11
#> 3 DOMAIN PH domain-like 2.7e-42    33   180 ENSMUST00000110663.8 
#> 4 DOMAIN PH domain-like 4.7e-42    33   180 ENSMUST00000110664.8 
#> 5 DOMAIN PH domain-like 3.9e-43    33   180 ENSMUST00000160134.7 
#> 6 DOMAIN PH domain-like 3.8e-43    33   180 ENSMUST00000161040.7 
#> 7 DOMAIN PH domain-like 2.5e-42    33   180 ENSMUST00000161812.7 
#> 8 DOMAIN PH domain-like 2.6e-21    35   110 ENSMUST00000162094.2 

## All outputs can be assigned to a variable and manipulated further using other functions:
ase(factRsample) %>% dplyr::filter(AStype == "CE")
#> Error in ase(factRsample) %>% dplyr::filter(AStype == "CE"): could not find function "%>%"