Run factR workflow

This wrapper function will perform the core factR workflow which includes:

1: Building CDS structures

2: Predicting NMD-sensitive transcripts

3: Translating open-reading frames

4: Quantify sequence conservation of alternative exons

Usage

runfactR(
  object,
  NMD_threshold = 50,
  cons_db = "phastCons",
  cons_type = "exon",
  cons_padding = 0,
  verbose = TRUE
)

Arguments

object

factR object

NMD_threshold

Minimum distance between PTC and downstream exon-exon junction to trigger NMD (Default: 50)

cons_db

Database to extract sequence conservation. Can be "phastCons" (Default) or "phylop"

cons_type

Feature to quantify conservation. Can be one of the following:

"exon" : Sequence conservation of the entire exon

"flanks" : Conservation of sequences flanking exons

"upstream" : Conservation of sequences upstream of exons

"downstream" : Conservation of sequences downstream of exons

cons_padding

Additional width to pad the sequence by. For cons_type "exons" and "flanks", paddings will be added on both sides.

verbose

Whether to show pipeline messages (Default: TRUE)

Value

Update factRObject with additional data from runfactR. See links below for detailed description of each function in this wrapper.

See also

Examples

## Load sample factRObject
data(factRsample)

## Run factR pipeline
factRsample <- runfactR(factRsample)
#> 🡆  Building CDS information
#>     Searching for reference mRNAs in query
#>     22 reference mRNAs found and its CDS were assigned
#>     Building database of annotated ATG codons
#>     Selecting best ATG start codon for remaining transcripts and determining open-reading frame
#>     61 new CDSs constructed
#> 
#>     Summary: Out of 155 transcripts in `gtf`,
#>     83 transcript CDSs were built
#>     ℹ Updating transcript feature data
#> 🡆  Running transcript-level NMD prediction
#>     Predicting NMD sensitivities for 83 mRNAs
#>   |                                                  | 0 % ~calculating    |+                                                 | 1 % ~00s            |++                                                | 2 % ~00s            |++                                                | 4 % ~00s            |+++                                               | 5 % ~00s            |++++                                              | 6 % ~00s            |++++                                              | 7 % ~00s            |+++++                                             | 8 % ~00s            |+++++                                             | 10% ~00s            |++++++                                            | 11% ~00s            |+++++++                                           | 12% ~00s            |+++++++                                           | 13% ~00s            |++++++++                                          | 14% ~00s            |++++++++                                          | 16% ~00s            |+++++++++                                         | 17% ~00s            |++++++++++                                        | 18% ~00s            |++++++++++                                        | 19% ~00s            |+++++++++++                                       | 20% ~00s            |+++++++++++                                       | 22% ~00s            |++++++++++++                                      | 23% ~00s            |+++++++++++++                                     | 24% ~00s            |+++++++++++++                                     | 25% ~00s            |++++++++++++++                                    | 27% ~00s            |++++++++++++++                                    | 28% ~00s            |+++++++++++++++                                   | 29% ~00s            |++++++++++++++++                                  | 30% ~00s            |++++++++++++++++                                  | 31% ~00s            |+++++++++++++++++                                 | 33% ~00s            |+++++++++++++++++                                 | 34% ~00s            |++++++++++++++++++                                | 35% ~00s            |+++++++++++++++++++                               | 36% ~00s            |+++++++++++++++++++                               | 37% ~00s            |++++++++++++++++++++                              | 39% ~00s            |++++++++++++++++++++                              | 40% ~00s            |+++++++++++++++++++++                             | 41% ~00s            |++++++++++++++++++++++                            | 42% ~00s            |++++++++++++++++++++++                            | 43% ~00s            |+++++++++++++++++++++++                           | 45% ~00s            |+++++++++++++++++++++++                           | 46% ~00s            |++++++++++++++++++++++++                          | 47% ~00s            |+++++++++++++++++++++++++                         | 48% ~00s            |+++++++++++++++++++++++++                         | 49% ~00s            |++++++++++++++++++++++++++                        | 51% ~00s            |++++++++++++++++++++++++++                        | 52% ~00s            |+++++++++++++++++++++++++++                       | 53% ~00s            |++++++++++++++++++++++++++++                      | 54% ~00s            |++++++++++++++++++++++++++++                      | 55% ~00s            |+++++++++++++++++++++++++++++                     | 57% ~00s            |+++++++++++++++++++++++++++++                     | 58% ~00s            |++++++++++++++++++++++++++++++                    | 59% ~00s            |+++++++++++++++++++++++++++++++                   | 60% ~00s            |+++++++++++++++++++++++++++++++                   | 61% ~00s            |++++++++++++++++++++++++++++++++                  | 63% ~00s            |++++++++++++++++++++++++++++++++                  | 64% ~00s            |+++++++++++++++++++++++++++++++++                 | 65% ~00s            |++++++++++++++++++++++++++++++++++                | 66% ~00s            |++++++++++++++++++++++++++++++++++                | 67% ~00s            |+++++++++++++++++++++++++++++++++++               | 69% ~00s            |+++++++++++++++++++++++++++++++++++               | 70% ~00s            |++++++++++++++++++++++++++++++++++++              | 71% ~00s            |+++++++++++++++++++++++++++++++++++++             | 72% ~00s            |+++++++++++++++++++++++++++++++++++++             | 73% ~00s            |++++++++++++++++++++++++++++++++++++++            | 75% ~00s            |++++++++++++++++++++++++++++++++++++++            | 76% ~00s            |+++++++++++++++++++++++++++++++++++++++           | 77% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 78% ~00s            |++++++++++++++++++++++++++++++++++++++++          | 80% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 81% ~00s            |+++++++++++++++++++++++++++++++++++++++++         | 82% ~00s            |++++++++++++++++++++++++++++++++++++++++++        | 83% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 84% ~00s            |+++++++++++++++++++++++++++++++++++++++++++       | 86% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 87% ~00s            |++++++++++++++++++++++++++++++++++++++++++++      | 88% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++     | 89% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 90% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++    | 92% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 93% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++   | 94% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++  | 95% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 96% ~00s            |+++++++++++++++++++++++++++++++++++++++++++++++++ | 98% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 99% ~00s            |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed=00s  
#>     ℹ Updating transcript feature data
#> 🡆  Translating amino acid sequences
#>     ❗ 1 CDSs do not begin with ATG
#>     ❗ 2 CDSs contain internal stop codon. Truncating CDS sequence to retain ORF
#> 🡆  Running AS-NMD testing
#>     ℹ Getting best reference for AS-NMD testing
#>     ℹ Testing AS-NMD exons
#>     ℹ Updating AS feature data
#> 🡆  Quantifiying conservation scores
#>     ℹ Retrieving phastCons60way.UCSC.mm10 scores
#>     ℹ Quantifying `exon` conservation scores with 0 padding