factRObject Class

A container for storing and processing custom transcriptome.

Details

The factRObject is a representation of custom-assembled transcriptome data. Coordinate and metadata information from an input transcriptome (GTF) is parsed and grouped into 3 data Sets which corresponds to the levels of gene (genes), transcript (txs) and alternative splicing (ase). With the aid of a compatible reference genome and expression data, functional and comparative analyses can be performed to better understand the complexity of a transcriptome.

Constructor

factRObject can be easily constructed as such:

createfactRObject(gtf, reference)

Accessors

Interact with a factRObject (x) the following ways:

summary(x): Prints preview of factRObject

head(x): Prints the first 6 metadata entries of the active Set

tail(x): Prints the last 6 metadata entries of the active Set

dim(x): Prints the dimension of the active set (features x samples)

nrow(x): Returns the number of features of the active set

ncol(x): Returns the number of samples

A factRObject consists of 3 `Sets` of data: gene, transcript and AS. At any one time, one of these `Sets` will be made the active Set and some factR2 functions will be called only on this active Set. Below are some ways to check and change the active Set:

listSets(x): Lists Sets in object

activeSet(x): Returns the active Set of object

activeSet(x)<-: Change the active Set

Each `Set` of an factRObject hold rich information of the features that it contains. Below are some convenient wrappers to display the metadata in each `Set`.

features(x): Displays metadata of active Set

genes(x) or gns(x): Displays gene metadata

transcripts(x) or txs(x): Displays transcript metadata

ase(x): Displays alternative splicing events metadata

factR2 also contain functions to plot exon-level and domain-level structures as an interactive plot:

plotTranscripts(x, "gene of interest"): Plots transcripts

plotDomains(x, "gene of interest")): Plots domains

If transcript-level counts data is provided, additional accessors are available:

counts(x): View/extract counts data

samples(x)): Prints sample metadata

rowNames(x)): Print feature names of current Set

colNames(x)): Prints sample names

colNames(x))<-: Set new sample names

ident(x)): Prints default sample identity

ident(x))<-: Set new sample identity

See also